added admin seeder

+++ s_drzewiasta/app/Http/Controllers/BranchesController.php
@@ -5,13 +5,14 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Branch;
 use Illuminate\Support\Facades\Auth;
+
 class BranchesController extends Controller
 {
     public function getBranches()
     {
         $branches = Branch::all(['id', 'text', 'parent_id']);
         $is_admin = false;
-        if (Auth::user() !== null){
+        if (Auth::user() !== null) {
             $is_admin = Auth::user()->is_admin;
         }
         return view('struct', ['data' => json_encode($branches), 'is_admin' => $is_admin]);
@@ -19,10 +20,13 @@ class BranchesController extends Controller
 
     public function updateBranches(Request $request)
     {
-        $data = json_decode($request->get('data'), true);
-//        $syy=[['data'=>'asd', 'asd'=>'loc'],['data'=>'ddff', 'asd'=> 'fafaf']];
-      var_dump($data);
-        Branch::truncate();
-        foreach($data as $arr)Branch::create($arr);
+        if (Auth::user() !== null) {
+            $is_admin = Auth::user()->is_admin;
+        }
+        if ($is_admin == true) {
+            $data = json_decode($request->get('data'), true);
+            Branch::truncate();
+            foreach ($data as $arr) Branch::create($arr);
+        }
     }
 }
